Add actions
Build/test/lint/coverage
