+++ .clang-lint-ignore
@@ -1,0 +1,1 @@
+# ignore third_party code from clang-format checks
