--- ci/travis-osx.sh
@@ -1,4 +1,0 @@
-#!/bin/bash
-
-brew unlink boost
-brew install ./boost@1.66--1.66.0.catalina.bottle.tar.gz
