+++ docs/conf.py.in
@@ -1,0 +1,72 @@
+# Configuration file for the Sphinx documentation builder.
+#
+# This file only contains a selection of the most common options. For a full
+# list see the documentation:
+# https://www.sphinx-doc.org/en/master/usage/configuration.html
+
+# -- Project information -----------------------------------------------------
+
+project = "@PROJECT_NAME@"
+copyright = "Apache-2.0"
+version = "@PROJECT_VERSION@"
+
+
+# -- General configuration ---------------------------------------------------
+
+# Add any Sphinx extension module names here, as strings. They can be
+# extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
+# ones.
+extensions = [
+    'breathe',
+    'exhale',
+    'sphinx.ext.githubpages',
+]
+
+# List of patterns, relative to source directory, that match files and
+# directories to ignore when looking for source files.
+# This pattern also affects html_static_path and html_extra_path.
+exclude_patterns = ['build', 'Thumbs.db', '.DS_Store']
+
+
+# -- Options for HTML output -------------------------------------------------
+
+# The theme to use for HTML and HTML Help pages.
+# See the documentation for a list of builtin themes.
+html_theme = 'furo'
+html_theme_options = {
+    "light_css_variables": {
+        "color-api-overall": "#cccccc",
+        "color-api-keyword": "#e62569",
+        "color-api-name": "#8ee22c",
+        "color-brand-content": "#67d8ef",
+    },
+}
+
+# Title of the docs
+html_title = "@PROJECT_NAME@"
+
+# Setup the breathe extension
+breathe_projects = {
+    "Doxygen XML": "@DOXYGEN_OUTPUT_DIR@/xml"
+}
+breathe_default_project = "Doxygen XML"
+
+# Setup the exhale extension
+exhale_args = {
+    "containmentFolder":      "@EXHALE_OUTPUT_DIR@",
+    "doxygenStripFromPath":   "@DOXYGEN_STRIP_PATH@",
+    "rootFileName":           "library_root.rst",
+    "rootFileTitle":          "Library API",
+    "fullApiSubSectionTitle": "Reference",
+    "createTreeView":         True,
+    "exhaleExecutesDoxygen":  True,
+    "exhaleUseDoxyfile":      True,
+    "contentsDirectives":     False,
+    "fullToctreeMaxDepth":    1,
+}
+
+# Tell sphinx what the primary language being documented is.
+primary_domain = 'cpp'
+
+# Tell sphinx what the pygments highlight language should be.
+highlight_language = 'cpp'
