+++ docs/genindex.rst.in
@@ -1,0 +1,2 @@
+Index
+=====
