+++ include/stream-client/stream/impl/http_socket.ipp
