--- streamclientConfig.cmake.in
@@ -1,18 +1,0 @@
-@PACKAGE_INIT@
-
-include(CMakeFindDependencyMacro)
-
-if (APPLE AND NOT DEFINED ENV{OPENSSL_ROOT_DIR})
-    find_program(HOMEBREW brew)
-    if (NOT HOMEBREW STREQUAL "HOMEBREW-NOTFOUND")
-        execute_process(COMMAND brew --prefix openssl
-            OUTPUT_VARIABLE OPENSSL_ROOT_DIR
-            OUTPUT_STRIP_TRAILING_WHITESPACE)
-    endif()
-endif()
-find_dependency(Threads)
-find_dependency(Boost 1.58)
-find_dependency(OpenSSL 1.0)
-
-include(${CMAKE_CURRENT_LIST_DIR}/streamclientTargets.cmake)
-check_required_components("@PROJECT_NAME@")
